+++ autoconfig/src/main/java/es/urjc/etsii/grafo/autoconfig/fill/FModeParam.java
@@ -6,14 +6,15 @@ import java.util.Set;
 
 public class FModeParam extends ParameterProvider {
 
-    private static final Set<String> generateValuesForNames = Set.of("fmode", "mode", "fMode");
+    private static final Set<String> generateValuesForNames = Set.of("fmode", "mode", "ofmode");
     @Override
     public boolean provides(Class<?> type, String paramName) {
-        return generateValuesForNames.contains(paramName);
+        return generateValuesForNames.contains(paramName.toLowerCase());
     }
 
     @Override
     public Object getValue(Class<?> type, String paramName) {
+        paramName = paramName.toLowerCase();
         if (!generateValuesForNames.contains(paramName)) {
             throw new IllegalArgumentException("Fmode only for parameter names %s, given %s".formatted(generateValuesForNames, paramName));
         }